requirements_linux.txt
# install torch with cuda
openai
einops
pydot
pyvis
astor
networkx
triton
wandb
transformers
datasets
evaluate
accelerate
causal-conv1d
mamba-ssm
flash-attn
torchtune
boto3
markdown
pyflowchart
streamlit_navigation_bar
streamlit_markmap
streamlit_timeline
streamlit-flow-component
arxiv
anthropic
langchain_anthropic
# paperswithcode-client
feedparser
cohere
pinecone[gRPC]
langchain-experimental
pypdf
cairosvg
python-weather
git+https://github.com/chengjunyan1/lm-evaluation-harness.git
git+https://github.com/felipemaiapolo/tinyBenchmarks
#git+https://YOUR_GITHUB_NAME:YOUR_GITHUB_TOKEN@github.com/allenai/exec_utils.git